Dockerizing the application that generates Swagger 1.2 specs
This commit Dockerizes the application by adding the Dockerfile,
the build.sh and the update to the Gruntfile that uses localhost.
In order to work, the application needs to bind to 0.0.0.0 to be
able to be accessed from outside the container. The discussion
to get it up and running is at the following URL:

https://github.com/wordnik/swagger-editor/issues/72

The application bounds to port 9000, so just adjust the docker
run command to properly bind to it.

docker run -it -p 9000:9000 marcellodesales/swagger-editor

*	new file:   Dockerfile
- The docker file that depends on the dependencies image.

*	modified:   Gruntfile.js
- Updating the binding of the ip address to the 0.0.0.0.

*	new file:   build.sh
- Just the build script to build the image independently from any
  other application.